# docs/swagger.yaml
basePath: /
definitions:
  models.KnowledgeArticle:
    properties:
      author:
        $ref: '#/definitions/models.User'
      authorID:
        type: string
      category:
        description: Simplified for now, mapped to enum in DB
        type: string
      content:
        type: string
      createdAt:
        type: string
      helpfulCount:
        type: integer
      id:
        type: string
      isVerified:
        type: boolean
      title:
        type: string
      updatedAt:
        type: string
      viewsCount:
        type: integer
    type: object
  models.User:
    properties:
      avatarURL:
        type: string
      createdAt:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      passwordHash:
        description: Nullable for LDAP users
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      updatedAt:
        type: string
    type: object
  models.UserRole:
    enum:
    - CONSUMER
    - STAFF
    - MANAGER
    type: string
    x-enum-varnames:
    - RoleConsumer
    - RoleStaff
    - RoleManager
  notification.SubscribeRequest:
    properties:
      endpoint:
        type: string
      keys:
        properties:
          auth:
            type: string
          p256dh:
            type: string
        type: object
    type: object
host: localhost:8080
info:
  contact:
    name: IT Support
  description: IT Broadcast Operations & Helpdesk System API. Provides endpoints for
    ticket management, knowledge base (Big Book), shift scheduling, and real-time
    notifications.
  license:
    name: MIT
  title: G-Report API
  version: "1.0"
paths:
...
  /consumer:
    get:
      description: Display consumer dashboard with recent tickets
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Consumer dashboard
      tags:
      - Consumer
  /consumer/articles/{id}:
    get:
      description: Display knowledge article details and increment view count
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
        "404":
          description: Article not found
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: View article detail
      tags:
      - Consumer
  /consumer/bigbook:
    get:
      description: Search Big Book articles with Redis caching
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.KnowledgeArticle'
            type: array
      security:
      - CookieAuth: []
      summary: Search knowledge base
      tags:
      - Consumer
  /consumer/ticket:
    post:
      consumes:
      - multipart/form-data
      description: Submit a new support ticket
      parameters:
      - description: Location
        in: formData
        name: location
        required: true
        type: string
      - description: Category
        in: formData
        name: category
        required: true
        type: string
      - description: Subject
        in: formData
        name: subject
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        required: true
        type: string
      - description: Urgency
        in: formData
        name: urgency
        type: string
      - description: Proof image
        in: formData
        name: proof_image
        type: file
      produces:
      - text/html
      responses:
        "302":
          description: Redirect to dashboard
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: object
      security:
      - CookieAuth: []
      summary: Create new ticket
      tags:
      - Consumer
  /consumer/tickets/{id}/details:
    get:
      description: Get ticket details and chat history as JSON
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ticket and activities
          schema:
            type: object
        "404":
          description: Ticket not found
          schema:
            type: object
      security:
      - CookieAuth: []
      summary: Get ticket details
      tags:
      - Consumer
  /consumer/tickets/{id}/reply:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Add a reply message to a ticket
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      - description: Reply message
        in: formData
        name: message
        required: true
        type: string
      produces:
      - text/html
      responses:
        "302":
          description: Redirect to dashboard
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Reply to ticket
      tags:
      - Consumer
  /consumer/tickets/{id}/stream:
    get:
      description: Server-Sent Events stream for real-time chat updates
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Ticket chat SSE stream
      tags:
      - Consumer
...
  /notifications/subscribe:
    post:
      consumes:
      - application/json
      description: Register push notification subscription
      parameters:
      - description: Push subscription
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/notification.SubscribeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Subscribed successfully
          schema:
            type: object
        "400":
          description: Invalid request
          schema:
            type: object
      security:
      - CookieAuth: []
      summary: Subscribe to push notifications
      tags:
      - Notifications
  /notifications/test:
    post:
      description: Send a test push notification to current user
      produces:
      - application/json
      responses:
        "200":
          description: Notification sent
          schema:
            type: object
        "500":
          description: Failed to send
          schema:
            type: object
      security:
      - CookieAuth: []
      summary: Send test notification
      tags:
      - Notifications
  /notifications/unsubscribe:
    post:
      consumes:
      - application/json
      description: Remove push notification subscription
      parameters:
      - description: Subscription endpoint
        in: body
        name: subscription
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Unsubscribed
          schema:
            type: object
        "400":
          description: Invalid request
          schema:
            type: object
      security:
      - CookieAuth: []
      summary: Unsubscribe from push notifications
      tags:
      - Notifications
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: user_id
    type: apiKey
swagger: "2.0"